Remove superfluous test class hierarchy
